requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework>=3.9.4
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.0